Daily treasury long-term rates

Treasury ceased publication of the 30-year constant maturity series on February 18, 2002 and resumed that series on February 9, 2006. To estimate a 30-year rate during that time frame, this series includes the Treasury 20-year Constant Maturity rate and an "adjustment factor," which may be added to the 20-year rate to estimate a 30-year rate during the period of time in which Treasury did not issue the 30-year bonds.

Usage

tr_long_term_rate(date = NULL)

Source

https://home.treasury.gov/treasury-daily-interest-rate-xml-feed

Arguments

date

character(1) or numeric(1) date in format yyyy or yyyymm. If NULL, all data is returned. Default NULL.

Value

A data.frame() containing the rates or NULL when no entries were found.

See also

Examples

# \donttest{
# get data for a single month
tr_long_term_rate("202201")
#> # A tibble: 60 × 3
#>    date       rate_type      rate
#>    <date>     <chr>         <dbl>
#>  1 2022-01-03 20 year        2.05
#>  2 2022-01-03 over 10 years  2   
#>  3 2022-01-03 real rate     -0.43
#>  4 2022-01-04 20 year        2.1 
#>  5 2022-01-04 over 10 years  2.06
#>  6 2022-01-04 real rate     -0.35
#>  7 2022-01-05 20 year        2.12
#>  8 2022-01-05 over 10 years  2.08
#>  9 2022-01-05 real rate     -0.27
#> 10 2022-01-06 20 year        2.12
#> # ℹ 50 more rows
# or for the entire year
tr_long_term_rate(2022)
#> # A tibble: 747 × 3
#>    date       rate_type      rate
#>    <date>     <chr>         <dbl>
#>  1 2022-01-03 20 year        2.05
#>  2 2022-01-03 over 10 years  2   
#>  3 2022-01-03 real rate     -0.43
#>  4 2022-01-04 20 year        2.1 
#>  5 2022-01-04 over 10 years  2.06
#>  6 2022-01-04 real rate     -0.35
#>  7 2022-01-05 20 year        2.12
#>  8 2022-01-05 over 10 years  2.08
#>  9 2022-01-05 real rate     -0.27
#> 10 2022-01-06 20 year        2.12
#> # ℹ 737 more rows
# }